renamed espresso tests files to follow coding convention
