+++ app/src/androidTest/java/org/liberty/android/fantastischmemo/ui/NewMCHintEspresso.java
@@ -31,7 +31,7 @@ import static org.hamcrest.Matchers.is;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class NewHintMCEspresso {
+public class NewMCHintEspresso {
 
     @Rule
     public ActivityTestRule<AnyMemo> mActivityTestRule = new ActivityTestRule<>(AnyMemo.class);
