+++ app/src/androidTest/java/org/liberty/android/fantastischmemo/ui/WorkoutEspresso.java
@@ -35,7 +35,7 @@ import static org.hamcrest.Matchers.is;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class WorkoutTest {
+public class WorkoutEspresso {
 
     @Rule
     public ActivityTestRule<AnyMemo> mActivityTestRule = new ActivityTestRule<>(AnyMemo.class);
